Decode custom Perseus entities to unicode
There are quite a few XML entities (&entities;) which Perseus uses
in its lexica which aren't in the HTML entity set. They provide a
.ent file describing what should be printed when they are
encountered, which we download and use to print useful characters
instead of entities.

For example, the following query had several '&ibreve;' entities,
which are now decoded:
  curl -d 'do=parse&lang=lat&q=Salaminii' http://localhost:8888/Perseus.cgi

In some cases the character doesn't have a unicode
representation, in which case a string like [xchar] is printed,
which is arguably slightly more readable than &xchar;.