Improvements to the bibliography lookup; most things working now
Currently Suetonius URLs still fail; need to debug that.
Also see Browser.pm for a server crashing bug that needs fixing.
Also need to test this with the Greek corpus.
origjump needs to be taken out once this is finished, too.
Then need to get Peter to sanity check it.