Break createFirstWindow out from createWindow
The logic specific to the startup window goes into createFirstWindow and
createWindow is used either for the first or subsequent windows.  We
only save one window state, which refers to the last created or
changed window.  This stops well short of full workspace management,
but works reasonably well as a simple solution.