This reverts Nick's commit c62062da79ddc7fce42481d491b927a54a4b3f4d
Nick's change did fix the reported bug of missing whitespace, but it
also added spurious whitespace elsewhere.  The problem is actually not
with the code but with Perseus' new version of the LSJ XML, which has
deleted whitespace where it should be and where it used to be.  It is
not immediately clear what heuristic to use to guess where to put it
back.  Since we may switch to using Helma's fork of the LSJ, which
does not have the whitespace bug, let's put this on hold for now.