Implement new jumpFrom feature: show lexicon again after JumpTo
We save the latest Perseus query within the main webpage and then if a
link from a lexicon entry sends us to a new main page of context, we
reload that same Perseus query again and show it in splitscreen.  The
correct way of doing this, of course, would be to rewrite the whole
interface to use AJAX also for the main window, so that both sides
could be updated independently of each other.