Fix bug where match was not found because it was just before range
We used to record only the end of a match, so for a word at the end of
a sentence, it was possible for the context to start at the next
sentence, and thus for the pattern not to be found in the context.
Now we also record the start of the match and work the context back
from there.