Use spawn rather than execFile to launch diogenes-server.pl
This is odd.  The parameter maxBuffer of execFile is supposed to apply
to STDOUT and STDERR of the child process (the server), but it clearly
also applies to the socket that the server uses to transmit the STDOUT
of the cgi script to the browser.  Large searches hit this limit and
Electron silently killed the server.  Raising maxBuffer fixed this,
but the better fix is to use spawn, which is not subject to maxBuffer.