Fixes for other corpora.
Documentary corpora do not work, so we throw an error.  They need code
to detect the start of a new document, and should be chunked by
document, though with line numbers added (as milestones?).  Many of
the symbols are undocumented.

Misc corpus does not work very well.  Does something odd with levels.
Coptic looks OK.