Revert to using GET rather than POST and up URL length limits.
This reverts 2d3fa9a81260fb5d0939aab1bc423c9c757c392d which was
designed to fix a bug when long query strings (e.g. with many forms
selected in a word-list search) are submitted.  But it turns out that
using POST for the form means that the back-button functionality
breaks in Electron (not in Chrome).  It's not clear whether this is
the correct behaviour or a bug, so we revert to GET and fix the
original bug by increasing the limits in the server.  It's not clear
what the limit is in Electron, but one Chrome dev message suggests
increasing it from 2M to 256M, so hopefully it is large enough.