Refactor diogenes-server to help solve Win32 issues
We now have a simpler, cleaner mechanism to run cgi scripts via
"require". After several experiments, we document the limitations of
the Windows fork emulation in Perl.  We add a timeout to Windows
connections, to permit a single child to service multiple clients.  We
can now interrupt long searches on Windows, and an error does not cause the
server to exit, which used to happen on Windows.

We tried using a new framework (Net::Server), but in Windows it did
not work any better, either for a forking server or a multiple
pre-forking server.