Fix strange windows bug that causes hanging when parsing
In some but not all Windows computers (but it does not depend on
version number), clicking on a word to parse it causes a lengthy hang
the first time it is done.  Not sure why, but it was slurping the
whole lexicon file and trying to parse it.  Somehow $/ was being
unset. Removing local fixes the bug; not sure why.