change shebang from /usr/bin/perl to /usr/bin/env
Generally using /usr/bin/perl is not a good shebang
because perl does not always live at /usr/bin/perl.
This is, for example, true for NixOS where multiple
versions of perl may be installed side by side in
multiple shells and for multiple users.
the correct way to lookup a program in the shebang
is /usr/bin/env perl.
Because here the shebang was /usr/bin/perl -w we
must use the -S switch on /usr/bin/env to pass
arguments to perl.
So the correct shebang is /usr/bin/env -S perl -w.

This applies to all kind of scripts. So also
/usr/bin/bash is not a good shebang. It should be
/usr/bin/env bash.