Fix multiple matching to find repetition of a word in a passage.
The old code presumed that the patterns were distinct and would not
work when you entered the same pattern multiple times.  Now the
repeat_matches flag indicates that all hits in a passage are to be
counted toward the total, rather than counting the number of patterns
that are hit in each passage, which is still the default behaviour.